feat: use hook to hide wallet connect and drag
